+++ pkg/services/move_documents/move_document_test.go
@@ -23,4 +23,3 @@ func TestMoveDocumentUpdaterServiceSuite(t *testing.T) {
 	}
 	suite.Run(t, hs)
 }
-
