--- pkg/services/move_documents/shared.go
@@ -1,1 +1,0 @@
-package movedocument
